Revert "Doctest errors corrected online"
This reverts commit 7456a26b

+++ pandas/core/arrays/sparse/array.py
@@ -1305,7 +1305,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [0.0, 0.0, 1.0, 2.0]
         Fill: 0.0
         IntIndex
-        Indices: array([2, 3], dtype=int32)
+        Indices: array([2, 3])
         """
         if is_dtype_equal(dtype, self._dtype):
             if not copy:
